% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcite.R
\name{gcite}
\alias{gcite}
\title{Google Citations Information}
\usage{
gcite(author, user, plot_wordcloud = TRUE, ...)
}
\arguments{
\item{author}{author name separated by spaces}

\item{user}{user ID for google Citations}

\item{plot_wordcloud}{should the wordcloud be plotted}

\item{...}{additional options passed to \code{\link{gcite_wordcloud}}}
}
\value{
List from either \code{\link{gcite_user_info}}
or \code{\link{gcite_author_info}}
}
\description{
Wraps getting the information from Google Citations and
plotting the wordcloud
}
