% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcite_paper_df.R
\name{gcite_paper_df}
\alias{gcite_paper_df}
\title{Get Paper Data Frame from Title URLs}
\usage{
gcite_paper_df(urls, verbose = TRUE, force = FALSE, ...)
}
\arguments{
\item{urls}{A character vector of urls, from 
\code{\link{all_papers$title_link}}}

\item{verbose}{Print diagnostic messages}

\item{force}{If passing a URL and there is a failure, should the 
program return \code{NULL}, passed to \code{\link{gcite_citation_page}}}

\item{...}{Additional arguments passed to \code{\link{GET}}}

\item{secure}{use https vs. http}
}
\value{
A \code{data.frame} of authors, journal, and citations
}
\description{
Get Paper Data Frame from Title URLs
}
