% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcite_stopwords.R
\name{gcite_stopwords}
\alias{gcite_stopwords}
\title{Google Cite Stopwords}
\usage{
gcite_stopwords()
}
\value{
Character Vector
}
\description{
Additional stopwords to remove from Google Cite results
}
\examples{
gcite_stopwords()
}
