% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcite_wordcloud.R
\name{gcite_wordcloud}
\alias{gcite_wordcloud}
\title{Wordcloud of Google Citations Information}
\usage{
gcite_wordcloud(paper_df, author_args = list(), title_args = list(),
  warn = FALSE)
}
\arguments{
\item{paper_df}{A \code{data.frame} with columns of authors and titles}

\item{author_args}{Arguments to pass to \code{\link{author_cloud}}}

\item{title_args}{Arguments to pass to \code{\link{title_cloud}}}

\item{warn}{should warnings be printed from wordcloud?}
}
\description{
Simple wrapper for \code{\link{author_cloud}}
and  \code{\link{title_cloud}}
}
