% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcite_wordcloud_spec.R
\name{gcite_wordcloud_spec}
\alias{gcite_wordcloud_spec}
\title{gcite Wordcloud default}
\usage{
gcite_wordcloud_spec(words, freq, min.freq = 1, max.words = Inf,
  random.order = FALSE, colors = c("#F768A1", "#DD3497", "#AE017E",
  "#7A0177", "#49006A"), vfont = c("sans serif", "plain"), ...)
}
\arguments{
\item{words}{words to be plotted}

\item{freq}{the frequency of those words}

\item{min.freq}{words with frequency below min.freq will not be plotted}

\item{max.words}{Maximum number of words to be plotted. least frequent terms dropped}

\item{random.order}{plot words in random order. If false, they will be plotted in decreasing frequency}

\item{colors}{color words from least to most frequent}

\item{vfont}{passed to text for the font}

\item{...}{additional options passed to \code{\link{wordcloud}}}
}
\value{
Nothing
}
\description{
Simple wrapper for \code{\link{wordcloud}} with 
different defaults
}
