% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_travis.R
\name{is_travis}
\alias{is_travis}
\title{Check if on Travis CI}
\usage{
is_travis()
}
\value{
Logical if user is named travis
}
\description{
Simple check for Travis CI for examples
}
\examples{
is_travis()
}
