% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cookies_txt.R
\name{set_cookies_txt}
\alias{set_cookies_txt}
\title{Set Cookies from Text file}
\usage{
set_cookies_txt(file)
}
\arguments{
\item{file}{tab-delimited text file of cookies, to be read in using
\code{\link[base]{readLines}}.  Comments should start the line with the pound
symbol}
}
\value{
Either \code{NULL} if no domains contain the word \code{"scholar"},
or an object of class \code{request} from \code{\link[httr]{set_cookies}}
}
\description{
Set Cookies from Text file
}
\note{
This function searches for domains that contain the word \code{"scholar"}
}
