% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcite_user_df.R, R/gcite_user_info.R
\name{gcite_author_info}
\alias{gcite_author_info}
\alias{gcite_author_info}
\alias{gcite_author_info}
\alias{gcite_author_info}
\title{Getting User Information of papers}
\usage{
gcite_author_info(author, verbose = TRUE, ask = TRUE, ...)

gcite_author_info(author, verbose = TRUE, ask = TRUE, ...)

gcite_author_info(author, verbose = TRUE, ask = TRUE, ...)

gcite_author_info(author, verbose = TRUE, ask = TRUE, ...)
}
\arguments{
\item{author}{author name separated by spaces}

\item{verbose}{Print diagnostic messages}

\item{ask}{If multiple authors are found, should a menu be given}

\item{...}{Not used}

\item{user}{user ID for google Citations}

\item{pagesize}{Size of pages, max 100, passed to \code{\link{gcite_url}}}

\item{verbose}{Verbose diagnostic printing}

\item{...}{arguments passed to \code{\link{gcite_user_info}}}

\item{user}{user ID for google Citations}

\item{pagesize}{Size of pages, max 100, passed to \code{\link{gcite_url}}}

\item{verbose}{Print diagnostic messages}

\item{...}{Not used}

\item{author}{author name separated by spaces}

\item{verbose}{Verbose diagnostic printing}

\item{ask}{If multiple authors are found, should a menu be given}

\item{...}{arguments passed to \code{\link{gcite_user_info}}}
}
\value{
A list of citations, citation indices, and a 
\code{data.frame} of authors, journal, and citations, and a 
\code{data.frame} of the links to all paper URLs.

A list of citations, citation indices, and a 
\code{data.frame} of authors, journal, and citations, and a 
\code{data.frame} of the links to all paper URLs.

A list of citations, citation indices, and a 
\code{data.frame} of authors, journal, and citations, and a 
\code{data.frame} of the links to all paper URLs.

A list of citations, citation indices, and a 
\code{data.frame} of authors, journal, and citations, and a 
\code{data.frame} of the links to all paper URLs.
}
\description{
Loops through pages for all information on Google Citations

Calls \code{\link{gcite_user_info}} after getting the user
identifier

Loops through pages for all information on Google Citations

Calls \code{\link{gcite_user_info}} after getting the user
identifier
}
\examples{
df = gcite_user_info(user = "T9eqZgMAAAAJ")
df = gcite_author_info(author = "John Muschelli")
df = gcite_user_info(user = "T9eqZgMAAAAJ")
df = gcite_author_info(author = "John Muschelli")
}
