% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcite_user_info.R
\name{gcite_user_info}
\alias{gcite_user_info}
\title{Getting User Information of papers}
\usage{
gcite_user_info(user, pagesize = 100, verbose = TRUE, secure = TRUE, ...)
}
\arguments{
\item{user}{user ID for google Citations}

\item{pagesize}{Size of pages, max 100, passed to \code{\link{gcite_url}}}

\item{verbose}{Print diagnostic messages}

\item{secure}{use https vs. http}

\item{...}{Not used}
}
\value{
A list of citations, citation indices, and a 
\code{data.frame} of authors, journal, and citations, and a 
\code{data.frame} of the links to all paper URLs and the character
string of the user name.
}
\description{
Loops through pages for all information on Google Citations
}
\examples{
if (!is_travis()) {
df = gcite_user_info(user = "T9eqZgMAAAAJ")
}
}
