% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcite_username.R
\name{gcite_username}
\alias{gcite_username}
\title{Google Citation Username Searcher}
\usage{
gcite_username(author, verbose = TRUE, ask = TRUE, secure = TRUE, ...)
}
\arguments{
\item{author}{author name separated by spaces}

\item{verbose}{Verbose diagnostic printing}

\item{ask}{If multiple authors are found, should a menu be given}

\item{secure}{use https vs. http}

\item{...}{arguments passed to \code{\link{GET}}}
}
\value{
A character vector of the username of the author
}
\description{
Search Google Citation for an author username
}
\examples{
if (!is_travis() & !is_cran()) {
gcite_username("John Muschelli")
}
}
